backend/blog/ws_middleware.py
from urllib.parse import parse_qs

from channels.db import database_sync_to_async
from django.contrib.auth import get_user_model
from django.contrib.auth.models import AnonymousUser

User = get_user_model()


@database_sync_to_async
def get_user_from_token(token):
    """Validate a JWT access token and return the corresponding User (or AnonymousUser)."""
    try:
        from rest_framework_simplejwt.tokens import UntypedToken
        UntypedToken(token)  # raises if invalid / expired
        import jwt as pyjwt
        from django.conf import settings
        decoded = pyjwt.decode(token, settings.SECRET_KEY, algorithms=["HS256"])
        user_id = decoded.get("user_id")
        return User.objects.get(id=user_id)
    except Exception:
        return AnonymousUser()


class JWTAuthMiddleware:
    """
    ASGI middleware that authenticates WebSocket connections via a JWT
    passed as a ?token=<access_token> query parameter.
    """

    def __init__(self, inner):
        self.inner = inner

    async def __call__(self, scope, receive, send):
        if scope["type"] == "websocket":
            query_string = scope.get("query_string", b"").decode()
            params = parse_qs(query_string)
            token_list = params.get("token", [])
            if token_list:
                scope["user"] = await get_user_from_token(token_list[0])
            else:
                scope["user"] = AnonymousUser()
        return await self.inner(scope, receive, send)
