backend/requirements.txt
Django>=6.0
djangorestframework
psycopg2-binary
django-environ
djangorestframework-simplejwt
django-cors-headers
bleach
Pillow
channels[daphne]
